feat(runtime): isolation, reload-context routing, manifest enforcement
- runtime-owned per-thread init-bundle-id stack replaces TLS (Rule 12);
  loaders behind RwLock; extension registry (register_extension)
- on_reload callbacks receive on_reload_user_data; retire-not-drop
  reload keeps pending slots invisible during the reload window
- manifest.id == fnv1a_64(name) now enforced at load via
  LoaderError::BundleTampered; validate() wired into the load path
- catch_unwind around plugin init (load + reload)
- raw-u32 dispatch_type validation, real function_count checks,
  double-destroy protection, runtime pointer nulled on destroy
- lua/js loaders: multi-contract registration, registerVtable carries
  contract version, free takes size/align; python: isolation nonce;
  dotnet: serde_json runtimeconfig, Box::leak removed

+++ crates/polyplug/tests/integration_ffi_null.rs
@@ -27,7 +27,7 @@ fn test_load_bundle_null_host() {
         unsafe { host_load_bundle(core::ptr::null(), path.as_ptr(), path.len()) };
     assert_eq!(
         result.code,
-        polyplug_abi::AbiErrorCode::InvalidPointer,
+        polyplug_abi::AbiErrorCode::InvalidPointer as u32,
         "load_bundle(null host) must return InvalidPointer"
     );
 }
@@ -42,7 +42,7 @@ fn test_load_bundle_null_path() {
         unsafe { ((*host).load_bundle)(host, core::ptr::null(), 0) };
     assert_eq!(
         result.code,
-        polyplug_abi::AbiErrorCode::InvalidPointer,
+        polyplug_abi::AbiErrorCode::InvalidPointer as u32,
         "load_bundle(null path) must return InvalidPointer"
     );
     // SAFETY: host is valid and was allocated by polyplug_runtime_create(core::ptr::null()).
